Support tmux escape sequence passthrough mode
Wrap OSC52 output in the Ptmux; escape sequence, taking advantage
of the tmux escape passthrough. This requires the tmux allow-passthrough
setting to be 'on'.

A new config option, 'tmux_passthrough' is added. It is set to false by
default for backwards compatibility. If this option is used, there is no
need to enable 'set-clipboard on'.

This option works well across ssh when both the client and server side
are running separate tmux sessions.