Fix ssh inside tmux
Issue:
If we use ssh inside tmux the TMUX env isn't passed over ssh so we use
the wrong osc52 code.
Fix:
if we check the TERM env instead that still works in local TMUX and also
in the remote case as well.
Notes:
tested locally and works in both cases.